refactor(runtime): `api/client` -> `api/host-api-resources-client`
